fix: Cancellled purchase kept promise dangling (#1504)
If the user does not proceed with the payment, the `RNIap.requestSubscription()` promise never resolves and remains dangling. It's also an issue if you want to show a "payment in progress" spinner while the user is interacting with the system level payment flows, as you don't get informed about the user cancelling, meaning, the spinner spins forever.  

Fixes: https://github.com/dooboolab/react-native-iap/issues/1479