fix: populate initial prices (#2582)
This pull request addresses two  issues:

**Amazon pricing**
When dealing with Amazon pricing,  consistently defaults
to 0.0. To rectify this, we've leveraged the 'original price' as the
source value for 'localizedPrice.' The primary objective of this patch
is to seamlessly populate both 'localizedPrice' and 'price' with values
derived from 'originalPrice.'

**purchaseUpdatedListener notifications**
During our investigation, we encountered a similar
[issue](#2487)
where notifications weren't being received within the
purchaseUpdatedListener. Our solution involved the reintroduction of the
startListening method.

---------

Co-authored-by: Zach <ztraianos@onair.events>