return ability to set 'obfuscatedAccountId' and 'obfuscatedProfileId' for android billing v3 (#1141)
Hey! I created this PR just because after I had bumped the version of lib to the new one, I noticed that the functionality of passing user id to android billing is missing. It seems that it was removed accidently and it is better to add it to fullfill API of android billing v3.
Also it seems that it related to this issue #1125